Correct integer length of printf

+++ src/ad7616_driver.c
@@ -575,7 +575,7 @@ void* DoDataAcquisition(void* vargp)
             // Open the previous file and append this sample line to it.  Always close the file to flush to disk.
             acquisitionFile = fopen(AcquisitionFilePath, "a");
             //fprintf(acquisitionFile, "%lu", ((nextticktime_ns-starttime_ns) / (1000*1000)));
-            fprintf(acquisitionFile, "%lu(%u)", ((now_ns-starttime_ns) / 1000), sizeof(unsigned long));
+            fprintf(acquisitionFile, "%llu(%u)", ((now_ns-starttime_ns) / 1000), sizeof(unsigned long));
             for (unsigned i = 0; i < SequenceSize; i++)
             {
                 fprintf(acquisitionFile, ",%d", separatedConversion[i]);
@@ -593,7 +593,7 @@ void* DoDataAcquisition(void* vargp)
             nextticktime_ns = nextticktime_ns + AcquisitiontPeriod_ns;
         }
 
-        unsigned long timeleftinperiod_ns = nextticktime_ns - now_ns;
+        unsigned long long timeleftinperiod_ns = nextticktime_ns - now_ns;
         // DIAGNOSTIC - Uncomment this line to get info on how much time is spent converting.
         // printf("Conversion time was %lu ns, sleeping %lu ns\n", (AcquisitiontPeriod_ns - timeleftinperiod_ns), timeleftinperiod_ns);
         usleep(timeleftinperiod_ns / 1000);
